+++ coq/workspace.mli
@@ -17,8 +17,10 @@
 
 module Flags : sig
   type t = private
-    { indices_matter : bool
-    ; impredicative_set : bool
+    { impredicative_set : bool
+    ; indices_matter : bool
+    ; type_in_type : bool
+    ; rewrite_rules : bool
     }
 end